Add proper copyright attributions as suggested by FSF.

+++ gnubg/board3d/shadow.c
@@ -1,13 +1,11 @@
 /*
- * shadow.c
- * by Jon Kinsey, 2003
+ * Copyright (C) 2003-2019 Jon Kinsey <jon_kinsey@hotmail.com>
+ * Copyright (C) 2003-2018 the AUTHORS
  *
- * 3d shadow functions
- *
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of version 3 or later of the GNU General Public License as
- * published by the Free Software Foundation.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -15,8 +13,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  *
  * $Id$
  */
@@ -57,7 +54,7 @@ shadowInit(BoardData3d * bd3d, renderdata * prd)
         g_print("No stencil buffer - no shadows\n");
         return;
     }
-    midStencilVal = 1u << (stencilBits - 1);
+    midStencilVal = (int) (1u << (stencilBits - 1));
     glClearStencil(midStencilVal);
 
     bd3d->shadowsInitialised = TRUE;
